[rename] withAuth 파일 이동
