+++ app/components/authToken/withAuth.tsx
